More comprehensive category related test
